refactor!: rename kftray-portforward to kftray-kube
After extracting kftray-ssl, kftray-hosts, and kftray-expose, the remaining crate is the Kubernetes orchestration layer: kubeconfig handling, pod watching, port-forward lifecycle, proxy pod deployment, TCP/UDP listeners, and the global registry. The kftray-kube name matches what it actually does.

Mechanical rename only:
- directory: crates/kftray-portforward -> crates/kftray-kube
- crate name: kftray-portforward -> kftray-kube
- module path: kftray_portforward -> kftray_kube
- dependents updated: kftray-tauri, kftui, kftray-network-monitor, kftray-mcp, kftray-expose
- frontend log filter strings updated
- CI workflow paths updated
