+++ crates/kftray-kube/src/.gitignore
