+++ crates/kftray-kube/src/kube/client/error.rs
