+++ crates/kftray-kube/src/kube/http_log_watcher.rs
