+++ crates/kftray-kube/src/kube/proxy.rs
