+++ crates/kftray-kube/src/registry.rs
