+++ crates/kftray-mcp/src/tools/portforward.rs
@@ -221,11 +221,11 @@ impl McpTool for StartPortForwardTool {
                         .await
                         .map_err(|e| e.to_string())
                     } else if protocol == "udp" {
-                        kftray_portforward::start_port_forward(configs, "udp")
+                        kftray_kube::start_port_forward(configs, "udp")
                             .await
                             .map_err(|e| e.to_string())
                     } else {
-                        kftray_portforward::start_port_forward(configs, "tcp")
+                        kftray_kube::start_port_forward(configs, "tcp")
                             .await
                             .map_err(|e| e.to_string())
                     };
@@ -347,11 +347,11 @@ impl McpTool for StartPortForwardTool {
                 .await
                 .map_err(|e| e.to_string())
             } else if protocol == "udp" {
-                kftray_portforward::start_port_forward(configs, "udp")
+                kftray_kube::start_port_forward(configs, "udp")
                     .await
                     .map_err(|e| e.to_string())
             } else {
-                kftray_portforward::start_port_forward(configs, "tcp")
+                kftray_kube::start_port_forward(configs, "tcp")
                     .await
                     .map_err(|e| e.to_string())
             };
@@ -467,10 +467,10 @@ impl McpTool for StopPortForwardTool {
                     kftray_commons::utils::db_mode::DatabaseMode::File,
                 )
                 .await
-                .map_err(|e| kftray_portforward::PortForwardError::Expose(e.to_string()))
+                .map_err(|e| kftray_kube::PortForwardError::Expose(e.to_string()))
             }
             _ => {
-                kftray_portforward::stop_port_forward(args.config_id.to_string()).await
+                kftray_kube::stop_port_forward(args.config_id.to_string()).await
             }
         };
 
@@ -520,7 +520,7 @@ impl McpTool for StopAllPortForwardsTool {
     }
 
     async fn execute(&self, _arguments: Option<Value>) -> CallToolResult {
-        match kftray_portforward::stop_all_port_forward().await {
+        match kftray_kube::stop_all_port_forward().await {
             Ok(responses) => {
                 let total_count = responses.len();
                 let stopped_count = responses.iter().filter(|r| r.status == 0).count();
