+++ crates/kftray-tauri/Cargo.toml
@@ -32,7 +32,7 @@ kftray-expose = { workspace = true }
 kftray-helper = { workspace = true }
 kftray-mcp = { workspace = true }
 kftray-network-monitor = { workspace = true }
-kftray-portforward = { workspace = true }
+kftray-kube = { workspace = true }
 kftray-ssl = { workspace = true }
 kftray-shortcuts = { workspace = true }
 kube = { workspace = true }
