+++ crates/kftui/Cargo.toml
@@ -27,7 +27,7 @@ futures = { workspace = true }
 kftray-commons = { workspace = true }
 kftray-expose = { workspace = true }
 kftray-network-monitor = { workspace = true }
-kftray-portforward = { workspace = true }
+kftray-kube = { workspace = true }
 kftray-ssl = { workspace = true }
 libc = { workspace = true }
 log = { workspace = true }
